Adds some initial V3 API support.
Since the CFAccessor has return types that depend on the V2 API, I have added v3 API methods to it instead of a V3 implementation.
This still allows different types of accessors (ex: caching) to implement the response of the V3 APIs but makes it easier to integrate alongside V2.

VITE-1251